Only cap the input if content length header is given
Currently RqLengthAware uses the InputStream#available to cap the input
if no Content-Length header is given, but the return value does not say
anything about the real content length, only how many bytes can be read
without blocking (what might be 0).

This now only returns a CapInputStream when a Content-Length is given in
the request and adds a testcase for the given case.