Do not allow closing of the live input stream
Currently the raw stream is passed to downstream consumers of RqLive,
this has the problem that such consumers can close the stream. In case
of a socket, this means the socket itself is closed and no response can
be send by takes.

This wraps the original live input stream into one that simply ignores
the close request.